// web/src/api/generated/public/models/ProductImage.ts
/* tslint:disable */
/* eslint-disable */
/**
 * HotShop API
 * Runtime-generated HTTP contract for HotShop
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 *
 * @export
 * @interface ProductImage
 */
export interface ProductImage {
    /**
     *
     * @type {string}
     * @memberof ProductImage
     */
    alt: string;
    /**
     *
     * @type {string}
     * @memberof ProductImage
     */
    url: string;
}

/**
 * Check if a given object implements the ProductImage interface.
 */
export function instanceOfProductImage(value: object): value is ProductImage {
    if (!('alt' in value) || value['alt'] === undefined) return false;
    if (!('url' in value) || value['url'] === undefined) return false;
    return true;
}

export function ProductImageFromJSON(json: any): ProductImage {
    return ProductImageFromJSONTyped(json, false);
}

export function ProductImageFromJSONTyped(json: any, ignoreDiscriminator: boolean): ProductImage {
    if (json == null) {
        return json;
    }
    return {

        'alt': json['alt'],
        'url': json['url'],
    };
}

export function ProductImageToJSON(json: any): ProductImage {
    return ProductImageToJSONTyped(json, false);
}

export function ProductImageToJSONTyped(value?: ProductImage | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {

        'alt': value['alt'],
        'url': value['url'],
    };
}