// web/src/api/generated/public/models/ProductPresentation.ts
/* tslint:disable */
/* eslint-disable */
/**
 * HotShop API
 * Runtime-generated HTTP contract for HotShop
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { ProductImage } from './ProductImage';
import {
    ProductImageFromJSON,
    ProductImageFromJSONTyped,
    ProductImageToJSON,
    ProductImageToJSONTyped,
} from './ProductImage';
import type { ProductSpecification } from './ProductSpecification';
import {
    ProductSpecificationFromJSON,
    ProductSpecificationFromJSONTyped,
    ProductSpecificationToJSON,
    ProductSpecificationToJSONTyped,
} from './ProductSpecification';

/**
 *
 * @export
 * @interface ProductPresentation
 */
export interface ProductPresentation {
    /**
     *
     * @type {string}
     * @memberof ProductPresentation
     */
    imageNote?: string;
    /**
     *
     * @type {Array<ProductImage>}
     * @memberof ProductPresentation
     */
    images?: Array<ProductImage>;
    /**
     *
     * @type {Array<ProductSpecification>}
     * @memberof ProductPresentation
     */
    specifications?: Array<ProductSpecification>;
}

/**
 * Check if a given object implements the ProductPresentation interface.
 */
export function instanceOfProductPresentation(value: object): value is ProductPresentation {
    return true;
}

export function ProductPresentationFromJSON(json: any): ProductPresentation {
    return ProductPresentationFromJSONTyped(json, false);
}

export function ProductPresentationFromJSONTyped(json: any, ignoreDiscriminator: boolean): ProductPresentation {
    if (json == null) {
        return json;
    }
    return {

        'imageNote': json['imageNote'] == null ? undefined : json['imageNote'],
        'images': json['images'] == null ? undefined : ((json['images'] as Array<any>).map(ProductImageFromJSON)),
        'specifications': json['specifications'] == null ? undefined : ((json['specifications'] as Array<any>).map(ProductSpecificationFromJSON)),
    };
}

export function ProductPresentationToJSON(json: any): ProductPresentation {
    return ProductPresentationToJSONTyped(json, false);
}

export function ProductPresentationToJSONTyped(value?: ProductPresentation | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {

        'imageNote': value['imageNote'],
        'images': value['images'] == null ? undefined : ((value['images'] as Array<any>).map(ProductImageToJSON)),
        'specifications': value['specifications'] == null ? undefined : ((value['specifications'] as Array<any>).map(ProductSpecificationToJSON)),
    };
}